fix: separate live logic from react renderer due to restriction in Next.js
